Fix resize issues with header logos and lines. Add JS.
Fixed the header logo and lines; now they can resize properly. Need to fix the mousover of the contact buttons.
Started to work on javascript interactions. Orange indicator lines added onmouseover of each card.
Also made tweaks to the card template. Still need to get demo and source buttons to rest at the button of the card and for the image to center between the bottom of the text and the top of the buttons.